+++ R/plugin.R
@@ -255,6 +255,12 @@ plugin_url <- function(platform) {
     return(url)
   }
 
+  if (os == "linux" && arch == "aarch64") {
+    # on linux arm download from our pre-built artifacts
+    url <- "https://github.com/r-xla/pjrt-builds/releases/download/pjrt/pjrt-6319f0d-linux-aarch64.tar.gz"
+    return(url)
+  }
+
   sprintf(
     "https://github.com/zml/pjrt-artifacts/releases/download/v%s/pjrt-%s_%s-%s.tar.gz",
     zml_version,
@@ -290,6 +296,8 @@ plugin_arch <- function() {
     return("amd64")
   } else if (Sys.info()["machine"] == "arm64") {
     return("arm64")
+  } else if (Sys.info()["machine"] == "aarch64") {
+    return("aarch64")
   } else if (.Platform$r_arch == "x64") {
     return("amd64")
   } else if (.Platform$r_arch == "arm64") {