+++ NEWS.md
@@ -1,5 +1,9 @@
 # pjrt (development version)
 
+## Platform Support
+
+* Added support for Linux ARM (aarch64) using CPU backend.
+
 ## Asynchronous API
 
 Operations such as host <-> device transfers and program execution were previously only